Add sharness as submodule
roundup turned out to be too limited.  For example, it was not obvious
how to break upon the first failing test.

So let's try sharness.  An additional advantage of sharness is that it
is compatible with git's native testing system; so it would be easy to
use the tests in core git later.