gc: Accelerate gc by using git rev-list --objects
Previously, gc was so slow that it was infeasible for a larger history.
This is fixed.  Instead of walking the commits and the trees, git is
told to output a stream of objects.  See [^stackoverflow] for details.

Note that the current code does not walk the reflog.  Therefore, silo
objects that are not directly reachable from the current history are
collected.

[^stackoverflow]: http://stackoverflow.com/questions/7348698/git-how-to-list-all-objects-in-the-database