fetch, push: Support remote.origin.silofetch and remote.origin.silopush to limit transfer
It might be useful to limit git silo transfer to a subset, for example
to small files.  It was not possible before to limit transfer when using
the default `git push` without additional parameters.  Now a number of
limiting paths can be stored in `remote.origin.silofetch` and
`remote.origin.silopush`.

I was wondering whether it would be more useful to specify include,
exclude regex patterns.  Since it wasn't obvious to me, I chose to
implement in this commit the most obvious solution: replace part of the
command line with values stored in the config.