getWorkingCopyEndl(): fix test for core.autocrlf
Previously, the test magically worked, although the syntax was
completely wrong.  If cfg_autocrlf was 'true', the shell executed the
command 'true', which always is true.  But if it contained something
else, the shell failed to run the command and thus took the else branch.
Now a real string comparison is used.