checkout --link: work around strange behavior of msysgit's bash
This adds a workaround for the problem described below.  When using
a hardlink during checkout, git-silo now ensures that the write
permissions are unset.  If it cannot fix the permissions, it falls back
to copying.

In msysgit's bash,

    touch a
    chmod a-w a
    ln a b
    rm b

unexpectedly gives 'a' write permissions.  Unlinking a files apparently
changes the permissions of the file.