fetch: Handle CTRL-C during fetch as expected
Previously, CTRL-C only interrupted the current transfer, but git-silo
continued fetching further files.

This is fixed.  scp errors are no longer ignored, and CTRL-C is
therefore handled properly.  To handle missing objects, a list of
available objects at the remote is built instead.  It is used to check
whether objects are available at the remote.  If not, a warning is
printed and the file is skipped.