filter_clean: Robustly handle CRLF checkouts
In case git had checked out a placeholder with CRLF, a subsequent git status
would have reported changes, because the placeholder in the worktree didn't
match the checked in placeholder.  placeholder might be checked out with CRLF
on Windows during merge conflicts in .gitattributes.  Due to the conflict, git
didn't see the right attributes and believed that the file would be a text file
and converted line endings to CRLF.

This is fixed.  filter_clean now removes CR from placeholder before returning
them to git.