Add attribute-based fetch and push filter
The new mechanism allows limiting fetch and push based on the
gitattribute `silo`.  The attribute `silo` can be set to
a comma-separated list of labels.  Per default, files that carry the
label `local` are excluded from transfer.  `--include` and `--exclude`
can be used to filter based on specific labels.  `--all` includes all
files.