checkout: Fix --link to link to same inode that is in silodir
Previously, --link didn't link if the path in the working copy has
a link count greater than 1.  But this doesn't reliably indicate that it
is linked to the file in silodir.  The path could also be linked to
a completely different file.

It's fixed.  git-silo now uses the inode number to determine whether it
needs to link.