checkout: Exit success if only 'local' content is missing
The gitattribute silo=local is useful for flagging content that should
not be pushed to central storage, such as less important files or
temporary files that can be recreated.  Tracking such files can be
useful, nonetheless, for example on a scratch file system.

To better support workflows that use silo=local, silo checkout is
changed to report a warning about missing content but exit with success.
It is expected that such objects might be missing, and callers of git
silo should not fail because of it.