checkout: Warn about missing gitattributes and fall back to placeholder
Using 'git mv' to rename a file might create sha1 placeholders without
the expected gitattributes.  Git is happy with the placeholder but gets
confused if it sees the real content, because it does not know that it
is supposed to filter it.  So it reports that the file has changed.

To avoid this situation, git silo checkout now refuses to checkout the
content if the expected gitattributes are missing.  Instead it prints
a warning and  keeps the sha1, so that git has nothing to complain
about.