checkout: Add heuristic to fight against Windows hardlink madness
Files on Windows need to be set to writable to delete a hardlink.
Quote from MSDN [^msdn]:

    For example if you reset the READONLY attribute on a hard link to
    delete that particular hard link, and there are multiple hard links
    to the actual file, then you will need to reset the READONLY bit on
    the file from one of the remaining hard links to bring the file and
    all remaining hard links back to the READONLY state.

Since there is no way to know the other hardlinks after deleting a file,
in practice, the file remains writable after deleting a hardlink.  `rm`
and emptying the Trash both handle it that way.

To contain this madness, try to set the silo object file to read-only
whenever the smudge filter sees it during checkout.  Report but ignore
errors.

[^msdn]: http://msdn.microsoft.com/en-us/library/windows/desktop/aa365006(v=vs.85).aspx