Preliminary port to SunOS
Most tests pass.  Obvious exceptions are packing, because 7z is not
available.  Some other tests might also fail.

Some commands are mapped to non-standard (GNU) tools to get the expected
behavior.  grep and egrep from `/usr/bin` in particular behave
strangely.

The ssh argument quoting seems to work differently on SunOS.  The
previous approach that sent a quoted string for `bash -c` did not work.
It complaint about unmatched single quotes.  Instead, the scripts now
are first written to a temporary file and then executed from there.
This requires a second ssh connection, which should not increase the
overall time, assuming that the transfer of the data dominates.

I tried dedup on real repos and it worked as expected.  git is a 32-bit
executable on our Sun machines and can only process files smaller than
2 GB.