add --attr, --no-attr to control whether gitattributes are modified
The wildcard setup has proved to be more robust in practice than the
automatic gitattributes handling.  When using wildcards, files can be
freely renamed as long as the extension is kept.  With explicit
gitattributes, renaming is problematic, because the gitattributes
usually are forgotton.

This introduced --no-attr to tell 'git add' to skip files if the
gitattributes are not already set.  The default is kept unchanged.