add: try to avoid computing sha1 by locating inode in silo.alternate
If an inode can be located in another silo object directory, its sha1
can be directly used.  For large files, this can be substantially faster
than computing a sha1 from the file content.

This changes `git silo add` to try to locate the inode at the paths that
are specified in the config `silo.alternate`.

The optimization can only be used with a patched git that supports
`update-index --assume-unchanged-once`.